File Module
FileUtil + tests
Writer + tests
