--- src/csvwriter/test.csv
@@ -1,3 +1,0 @@
-field1,field2,field3
-bob1,bob2,bob3
-bob10,bob20,bob30
