+++ src/files/test.csv
@@ -1,0 +1,3 @@
+field2,field1,field3
+bob2,bob1,bob3
+bob20,bob10,bob30
